improving app usage

+++ src/main/java/com/spring/medicalappointments/appointment/AppointmentService.java
@@ -6,12 +6,15 @@ import com.spring.medicalappointments.appointment.dto.AppointmentRequest;
 import com.spring.medicalappointments.appointment.dto.AppointmentResponse;
 import com.spring.medicalappointments.doctor.Doctor;
 import com.spring.medicalappointments.doctor.DoctorRepository;
+import com.spring.medicalappointments.exception.ApiNotFoundException;
+import com.spring.medicalappointments.exception.ApiRequestException;
 import com.spring.medicalappointments.patient.Patient;
 import com.spring.medicalappointments.patient.PatientRepository;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
+import java.time.LocalDateTime;
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
@@ -32,38 +35,62 @@ public class AppointmentService {
 
     public List<AppointmentResponse> getDoctorAppointments(Long doctorId) {
         if (!doctorRepository.existsById(doctorId)) {
-            throw new NoSuchElementException("Doctor: " +doctorId+ " not found");
+            throw new ApiNotFoundException("Doctor: " + doctorId + " not found");
         }
 
-        return appointmentRepository.findAllByDoctorId(doctorId).stream()
+        return appointmentRepository.findAllByDoctorId(doctorId)
+                .stream()
                 .map(appointmentDTOMapper)
                 .collect(Collectors.toList());
     }
-
 
     public List<AppointmentResponse> getPatientAppointments(Long patientId) {
         if (!patientRepository.existsById(patientId)) {
-            throw new NoSuchElementException("Patient: " +patientId+ " not found");
+            throw new ApiNotFoundException("Patient: " + patientId + " not found");
         }
 
-        return appointmentRepository.findAllByPatientId(patientId).stream()
+        return appointmentRepository.findAllByPatientId(patientId)
+                .stream()
                 .map(appointmentDTOMapper)
                 .collect(Collectors.toList());
     }
 
+    public List<AppointmentResponse> getAppointmentsByDoctorAndDate(Long doctorId, LocalDate date) {
+        if (!doctorRepository.existsById(doctorId)) {
+            throw new ApiNotFoundException("Doctor: " + doctorId + " not found");
+        }
+
+        return appointmentRepository.findAllByDoctorIdAndDate(doctorId, date)
+                .stream()
+                .map(appointmentDTOMapper)
+                .collect(Collectors.toList());
+    }
+
+    public List<AppointmentResponse> getAppointmentsByPatientAndDate(Long patientId, LocalDate date) {
+        if (!patientRepository.existsById(patientId)) {
+            throw new ApiNotFoundException("Patient: " + patientId + " not found");
+        }
+
+        return appointmentRepository.findAllByPatientIdAndDate(patientId, date)
+                .stream()
+                .map(appointmentDTOMapper)
+                .collect(Collectors.toList());
+    }
 
     public void createAppointment(AppointmentRequest appointmentRequest) {
+        Long patientId = appointmentRequest.patientId();
+        Long doctorId = appointmentRequest.doctorId();
+        LocalDateTime date = appointmentRequest.date();
 
-        /* checkear que no se pueda crear un appointment para un medico si ya tiene uno a esa hora :
-                List<Appointment> doctorAppointments = appointmentRepository.findAllByDoctorId(appointmentRequest.doctorId());
-        if (doctorAppointments.contains())*/
+        Patient patient = patientRepository.findById(patientId)
+                .orElseThrow(() -> new ApiNotFoundException("Patient: "+ patientId +" not found"));
 
-        Patient patient = patientRepository.findById(appointmentRequest.patientId())
-                .orElseThrow(() -> new NoSuchElementException("Patient not found"));
+        Doctor doctor = doctorRepository.findById(doctorId)
+                .orElseThrow(() -> new ApiNotFoundException("Doctor: "+ doctorId +" not found"));
 
-        Doctor doctor = doctorRepository.findById(appointmentRequest.doctorId())
-                .orElseThrow(() -> new NoSuchElementException("Doctor not found"));
-
+        if (existsByDoctorIdAndDateTime(doctorId, date)) {
+            throw new ApiRequestException("Doctor: " +doctorId+ " already have an appointment on: " + date);
+        }
 
         Appointment appointment = new Appointment(appointmentRequest.date(),
                 appointmentRequest.detail(), appointmentRequest.status(), doctor, patient);
@@ -71,12 +98,23 @@ public class AppointmentService {
         appointmentRepository.save(appointment);
     }
 
-
+    // cant update doctor or patient
     public void updateAppointment(Long id, AppointmentRequest appointmentRequest) {
         Appointment appointment = appointmentRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("Appointment: " + id + "NOT FOUND"));
+                .orElseThrow(() -> new ApiNotFoundException("Appointment: " + id + "NOT FOUND"));
 
-        appointment.setDate(appointmentRequest.date());
+        Long doctorId = appointmentRequest.doctorId();
+        LocalDateTime date = appointmentRequest.date();
+
+        if (doctorRepository.existsById(doctorId)) {
+            throw new ApiNotFoundException("Doctor: " +doctorId+ " doesnt exist");
+        }
+
+        if (existsByDoctorIdAndDateTime(doctorId, date)) {
+            throw new ApiRequestException("Doctor: " +doctorId+ " already have an appointment on: " + date);
+        }
+
+        appointment.setDateAndHour(date);
         appointment.setDetail(appointmentRequest.detail());
         appointment.setStatus(appointmentRequest.status());
 
@@ -85,7 +123,14 @@ public class AppointmentService {
 
 
     public void deleteAppointment(Long id) {
+        if (!appointmentRepository.existsById(id)) {
+            throw new ApiNotFoundException("Appointment: "+id+" not found");
+        }
         appointmentRepository.deleteById(id);
+    }
+
+    private boolean existsByDoctorIdAndDateTime(Long doctorId, LocalDateTime localDateTime) {
+        return appointmentRepository.existsByDoctorIdAndDateTime(doctorId, localDateTime);
     }
 
 }
