+++ src/main/java/com/spring/medicalappointments/patient/PatientService.java
@@ -1,12 +1,14 @@
 package com.spring.medicalappointments.patient;
 
+import com.spring.medicalappointments.exception.ApiNotFoundException;
+import com.spring.medicalappointments.exception.ApiRequestException;
 import com.spring.medicalappointments.patient.dto.PatientDTOMapper;
 import com.spring.medicalappointments.patient.dto.PatientRequest;
 import com.spring.medicalappointments.patient.dto.PatientResponse;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
-import java.util.NoSuchElementException;
 import java.util.stream.Collectors;
 
 @Service
@@ -30,37 +32,40 @@ public class PatientService {
     public PatientResponse getPatientById(Long id) {
         return patientRepository.findById(id)
                 .map(patientDTOMapper)
-                .orElseThrow(() -> new IllegalArgumentException("not found"));  // handle this exception propertly
+                .orElseThrow(() -> new ApiNotFoundException("Patient: "+id+" not found"));
     }
 
     public PatientResponse getPatientByName(String name) {
         return patientRepository.findByName(name)
                 .map(patientDTOMapper)
-                .orElseThrow(() -> new IllegalArgumentException("not found"));   // handle this exception propertly
+                .orElseThrow(() -> new ApiNotFoundException("Patient: "+name+" not found"));
     }
 
     public void addPatient(PatientRequest patientRequest) {
         if (patientRepository.findByEmail(patientRequest.email()).isPresent()) {
-            throw new IllegalStateException("A user with that email already exists. Try with another email");
+            throw new ApiRequestException("A user with that email already exists. Try with another email");
         }
 
         Patient patient = new Patient(patientRequest.name(), patientRequest.email(), patientRequest.password(), patientRequest.dob());
         patientRepository.save(patient);
     }
 
+    @Transactional
     public void updatePatient(Long id, PatientRequest patientRequest) {
+        Patient patient = patientRepository.findById(id)
+                .orElseThrow(() -> new ApiNotFoundException("Patient: "+id+" not found"));
 
-        Patient updatedPatient = patientRepository.findById(id)
-                .orElseThrow(() -> new NoSuchElementException("Patient: " + id + "NOT FOUND"));
-
-        updatedPatient.setName(patientRequest.name());
-        updatedPatient.setPassword(patientRequest.password());
-        updatedPatient.setEmail(patientRequest.email());
-        updatedPatient.setDob(patientRequest.dob());
-        patientRepository.save(updatedPatient);
+        patient.setName(patientRequest.name());
+        patient.setPassword(patientRequest.password());
+        patient.setEmail(patientRequest.email());
+        patient.setDob(patientRequest.dob());
+        patientRepository.save(patient);
     }
 
     public void deletePatient(Long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new ApiNotFoundException("Patient: "+id+" not found");
+        }
         patientRepository.deleteById(id);
     }
